ljc/stock_research/requirements.txt
akshare>=1.12.0
yfinance>=0.2.36
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.40.0
jinja2>=3.1.0
plotly>=5.18.0
lxml>=4.9.0
python-dotenv>=1.0.0
tqdm>=4.66.0
